Refactor product and store ID types from Int to Long
Updated ID data types for products and stores to align with Long, ensuring consistency across service, repository, and controller layers. Added new methods and endpoints for fetching products and stores by related entities, improving functionality and code maintainability.

+++ src/main/kotlin/com/example/project24/product/ProductRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ProductRepository : JpaRepository<Product, Int> {
+interface ProductRepository : JpaRepository<Product, Long> {
     @Query(
         value = "SELECT p FROM Product p WHERE p.store.id = :storeId"
     )
